split greenbotocore and s3 tools
